requirements.txt
django
django-tables2
django-select2
django-filter
django-crispy-forms
Pillow